% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releaseworkflow-package.R
\docType{package}
\name{releaseworkflow-package}
\alias{releaseworkflow}
\alias{releaseworkflow-package}
\title{releaseworkflow: What the Package Does (One Line, Title Case)}
\description{
For more information visit <https://2degrees-investing.org/>.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
